refactor: Overlay 컴포넌트 스타일 분리
